Draws grid
--> draws a grid
--> able to add obstacles
--> able to add goals

+++ source/projects/Movement/SteeringBehaviors/Flocking/Flock.cpp
@@ -409,7 +409,7 @@ Elite::Vector2 Flock::GetAverageNeighborPos() const
 	{
 		avarage += m_Neighbors[idx]->GetPosition();
 	}
-	avarage /= m_NrOfNeighbors;
+	avarage /= (float)m_NrOfNeighbors;
 
 
 	return avarage;
